feat: Phase 5 training loop (#9)

+++ crates/aarambh-ai-data/src/loader.rs
@@ -146,7 +146,7 @@ mod tests {
         }
 
         fn vocab_size(&self) -> usize {
-            self.vocab.len() as usize
+            self.vocab.len()
         }
 
         fn eos_token_id(&self) -> u32 {
